Add Bluetooth comms code.
This is just some initial basic communications code and is not
yet hooked in with the main code. I also haven't been able to fully
test it with a Bluetooth module due to some hardware issues. However,
based on the testing I have been able to perform I'm reasonably
confident that this mostly ought to work.